add weather service
This will be a convenient place to store code related to communicating
with the remote weather service.

I also added a new test for a new city to start forcing the design to
allow different addresses.

The http level integration tests are still fast enough that the need for
testing lower level interfaces is negligible.

I am using ActiveModel Validations on Address since they are there. This
is picked up by the rescue_responses system. I am kind of surprised that
it isn't rescued by default.