add forecasts
Users will be able to get a full forecast from the remote weather
service. This one is using a resourceful style since forecasts are
data objects and a list of data objects fits with an index action.

Also, move address into app/models/address.rb because it is getting
reused and it is a model.