add cache info to cache scenario
I went with a header to expose the caching information, as that is where
you should put control level information.

I always return the cache info as I think it would be better for the
user to know when the cache expires. I would assume that they would
avoid re-requesting until the expiration.

I debated using the Cache-Control header, but it has been a while and I don't want
to research if using it is a good idea. So, I went with a custom header.

ActiveSupport::Cache::Store has no public methods exposing the
expiration information, so I had to unfortunately use private methods to
access that.